Fix: normalize interfaces and pure functions in Filter component to avoid css resource load order incorrect
