+++ webapp/app/components/Filters/FilterControl.tsx
@@ -1,14 +1,15 @@
 import React, { Component, PureComponent, Suspense, ReactNode } from 'react'
 import {
-  ControlOptions,
   renderInputText,
   renderNumberRange,
   renderSelect,
   renderDate,
-  renderDateRange,
-  deserializeDefaultValue,
-  IControlBase
+  renderDateRange
 } from './'
+import { IControlBase, ControlOptions } from './types'
+import {
+  deserializeDefaultValue
+} from './util'
 import { FilterTypes } from './filterTypes'
 import { Form } from 'antd'
 import { WrappedFormUtils } from 'antd/lib/form/Form'
