+++ webapp/app/components/Filters/config/FilterForm.tsx
@@ -31,12 +31,12 @@ const RadioGroup = Radio.Group
 const RadioButton = Radio.Button
 
 import { FilterTypeList, FilterTypesLocale, FilterTypes } from '../filterTypes'
+import { renderDate } from '..'
+import { InteractionType } from '../types'
 import {
-  renderDate,
   getOperatorOptions,
-  getDatePickerFormatOptions,
-  InteractionType
-} from '..'
+  getDatePickerFormatOptions
+} from '../util'
 import DatePickerFormats, {
   DatePickerFormatsLocale,
   DatePickerDefaultValuesLocales,
