+++ webapp/app/components/Filters/config/RelatedInfoSelectors.tsx
@@ -21,7 +21,7 @@
 import React, { PureComponent } from 'react'
 import classnames from 'classnames'
 import { Form, Row, Col, Checkbox, Select, Radio, Empty } from 'antd'
-import { InteractionType, IControlRelatedField } from '..'
+import { InteractionType, IControlRelatedField } from '../types'
 import { RadioChangeEvent } from 'antd/lib/radio'
 import { CheckboxChangeEvent } from 'antd/lib/checkbox'
 import { IRelatedItemSource, IRelatedViewSource } from './FilterConfig'
