+++ webapp/app/utils/migrationRecorders/globalControl.ts
@@ -1,4 +1,4 @@
-import { IGlobalControl } from 'app/components/Filters'
+import { IGlobalControl } from 'app/components/Filters/types'
 import { DEFAULT_CACHE_EXPIRED } from 'app/globalConstants'
 import { IMigrationRecorder } from '.'
